+++ TwoHoSun/Application/Constants.swift
@@ -23,3 +23,31 @@ extension Constants {
         static let loginState = "loginState"
     }
 }
+
+extension Constants {
+
+    struct URLConst {
+        static var baseURL: String {
+            guard let baseURL = Bundle.main.infoDictionary?["BASE_URL"] as? String else {
+                fatalError("BASE_URL is missing in the Info.plist")
+            }
+            return baseURL.replacingOccurrences(of: " ", with: "")
+        }
+
+        static var cnetAPI: String {
+            return "http://www.career.go.kr/cnet/openapi/getOpenApi"
+        }
+    }
+}
+
+extension Constants {
+
+    struct APIKeyConst {
+        static var apiKey: String {
+            guard let key = Bundle.main.object(forInfoDictionaryKey: "SCHOOL_API_KEY") as? String else {
+                fatalError("SCHOOL_API_KEY error")
+            }
+            return key
+        }
+    }
+}
